@@ -17,7 +17,6 @@ classifiers = [
     "Environment :: Other Environment",
     "Intended Audience :: Developers",
     "Operating System :: OS Independent",
-    "Programming Language :: Python :: 3.12",
     "Topic :: Software Development :: Libraries :: Python Modules"
 ]
 packages = [
@@ -25,12 +24,11 @@ packages = [
     { include = "tests", format = "sdist" },
 ]
 include = [
-    "LICENSE", 
-    "CHANGELOG.rst",
+    { path = "CHANGELOG.rst", format = "sdist" }
 ]
 
 [tool.poetry.dependencies]
-python = "^3.8"
+python = "^3.9"
 protobuf = "^4.24.3"
 http-ece = "^1.1.0"
 cryptography = ">=2.5"
@@ -44,7 +42,6 @@ docs = ["sphinx", "sphinx-rtd-theme", "sphinx-autodoc-typehints"]
 
 [tool.poetry.group.dev.dependencies]
 pre-commit = "*"
-tox = "*"
 pytest = "*"
 pytest-mock = "*"
 pytest-asyncio = "*"
@@ -53,6 +50,8 @@ async-timeout = ">=3.0.0"
 mypy-protobuf = "^3.5.0"
 pytest-socket = "^0.7.0"
 aioresponses = "^0.7.6"
+types-protobuf = "^4"
+mypy = "^1"
 
 [tool.pytest.ini_options]
 testpaths = "tests"
@@ -71,11 +70,11 @@ requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.ruff]
-target-version = "py38"
+target-version = "py39"
 exclude = [
     "firebase_messaging/proto/*"
 ]
-select = [
+lint.select = [
   "E",  # pycodestyle
 #  "D",  # pydocstyle
   "F",  # pyflakes
@@ -85,15 +84,15 @@ select = [
   "I",  # isort
   "S",  # bandit
 ]
-ignore = [
+lint.ignore = [
   "D105",  # Missing docstring in magic method
   "D107",  # Missing docstring in `__init__`
 ]
 
-[tool.ruff.pydocstyle]
+[tool.ruff.lint.pydocstyle]
 convention = "pep257"
 
-[tool.ruff.per-file-ignores]
+[tool.ruff.lint.per-file-ignores]
 "tests/*.py" = [
     "D100",
     "D101",
@@ -112,6 +111,12 @@ convention = "pep257"
 [tool.mypy]
 exclude = [
     'tests/.*',  # TOML literal string (single-quotes, no escaping necessary)
-    'docs/.*'
+    'docs/.*',
+    'firebase_messaging/proto/.*'
 ]
 disallow_untyped_defs = true
+
+[tool.doc8]
+paths = ["docs"]
+ignore = ["D001"]
+ignore-path =  ["docs/build"]
